+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Development
   being checked against
 - Geo: Change function signatures to use latlon tuple instead of latitude and longitude
 - Geo: Enable querying station id instead of latlon within interpolate and summarize
+- Geo: Allow using values of nearby stations instead of interpolated values
 
 0.49.0 (28.11.2022)
 *******************